+++ soroban-sdk/src/tests.rs
@@ -5,6 +5,7 @@ mod contract_add_i32;
 mod contract_call_stack;
 mod contract_invoke;
 mod contract_invoker_account;
+mod contract_invoker_client;
 mod contract_udt_enum;
 mod contract_udt_struct;
 mod contract_udt_struct_tuple;
